Remove the listener on twitter profile card links
The personal link on a twitter profile has a different DOM
representation than links in tweets. In particular, the url
scheme is missing entirely. Also, when links are truncated
with a "…", the full URL will not be present.

For these reasons, attaching the click listener to this
specific type of link is buggy more often than not, and has
been removed.